requirements.txt
duckduckgo2==0.242

